refactor: use repositories for the data layer
