+++ CHECKS
@@ -1,3 +1,3 @@
 ATTEMPTS=3
 
-/checks Notion to Anki
+/api/checks 2anki.net
