--- src/lib/User/comparePassword.ts
@@ -1,8 +1,0 @@
-import bcrypt from 'bcryptjs';
-
-export default function comparePassword(
-  password: string,
-  hash: string
-): boolean {
-  return bcrypt.compareSync(password, hash);
-}
