--- src/lib/User/hashPassword.ts
@@ -1,5 +1,0 @@
-import bcrypt from 'bcryptjs';
-
-export default function hashPassword(password: string): string {
-  return bcrypt.hashSync(password, 12);
-}
