--- src/lib/favorite/all.ts
@@ -1,7 +1,0 @@
-import DB from '../storage/db';
-
-export default function all(owner: number) {
-  return DB('favorites').select('*').where({
-    owner,
-  });
-}
