--- src/lib/favorite/create.ts
@@ -1,9 +1,0 @@
-import DB from '../storage/db';
-
-export default function create(id: string, owner: number, type: string) {
-  return DB('favorites').insert({
-    object_id: id,
-    type,
-    owner,
-  });
-}
