+++ src/routes/VersionRouter.ts
@@ -1,0 +1,10 @@
+import express from 'express';
+import VersionController from '../controllers/VersionController';
+
+const VersionRouter = () => {
+  const router = express.Router();
+  router.get('/api/version', VersionController.getVersionInfo);
+  return router;
+};
+
+export default VersionRouter;
