--- src/routes/upload/purgeBlockCache.ts
@@ -1,4 +1,0 @@
-import DB from '../../lib/storage/db';
-
-export const purgeBlockCache = (owner: string) =>
-  DB('blocks').del().where({ owner });
