+++ src/seeds/01_users.ts
@@ -1,17 +1,20 @@
 import { Knex } from 'knex';
 
-import hashPassword from '../lib/User/hashPassword';
+import UsersRepository from '../data_layer/UsersRepository';
+import DB from '../lib/storage/db';
 
 export async function seed(knex: Knex): Promise<void> {
   // Deletes ALL existing entries
   await knex('users').del();
+
+  const userRepostiory = new UsersRepository(DB);
 
   // Inserts seed entries
   await knex('users').insert([
     {
       id: 21,
       name: 'Alexander Alemayhu',
-      password: hashPassword('ichiban'),
+      password: userRepostiory.getHashPassword('ichiban'),
       email: 'alexander@alemayhu.com',
     },
   ]);
